16.10web

+++ installer/bin/check.sh
@@ -19,20 +19,23 @@ touch /var/run/check.lock
 myUPTIME=$(awk '{print int($1/60)}' /proc/uptime)
 for i in $myIMAGES
   do
-    myCIDSTATUS=$(docker exec $i supervisorctl status)
-      if [ $? -ne 0 ];
-        then
-          myCIDSTATUS=1
-        else
-          myCIDSTATUS=$(echo $myCIDSTATUS | egrep -c "(STOPPED|FATAL)")
-      fi
-      if [ $myUPTIME -gt 4 ] && [ $myCIDSTATUS -gt 0 ];
-        then
-          echo "Restarting "$i"."
-          systemctl stop $i
-          sleep 5
-          systemctl start $i
-      fi
+    if [ "$i" != "ui-for-docker" ] && [ "$i" != "netdata" ];
+      then
+        myCIDSTATUS=$(docker exec $i supervisorctl status)
+        if [ $? -ne 0 ];
+          then
+            myCIDSTATUS=1
+          else
+            myCIDSTATUS=$(echo $myCIDSTATUS | egrep -c "(STOPPED|FATAL)")
+        fi
+        if [ $myUPTIME -gt 4 ] && [ $myCIDSTATUS -gt 0 ];
+          then
+            echo "Restarting "$i"."
+            systemctl stop $i
+            sleep 5
+            systemctl start $i
+        fi
+    fi
 done
 
 rm /var/run/check.lock
